requirements.txt
Django==3.2.15
flake8==5.0.4
django-environ==0.9.0
django-extensions==3.2.1
Pillow==9.2.0
ipdb==0.13.9
Faker==15.0.0
weasyprint==56.1
django-lifecycle==1.0.0
celery[redis]==5.2.0
django-celery-results==2.2.0
django-celery-beat==2.2.1
requests==2.28.1
pymemcache==4.0.0
django-silk==5.0.2
django-debug-toolbar==3.7.0
pytest==7.2.0
pytest-django==4.5.2
pytest-mock==3.10.0
coverage==6.5.0
gunicorn==20.1.0
pytest-factoryboy==2.5.0
beautifulsoup4==4.11.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
drf-yasg==1.21.4